refactor: Pawn이 색깔 상수와 기본 생성자를 가진다. - 기본생성자 사용시 color는 "white"가 된다. - Pawn.WHITE_COLOR = "white"; - Pawn.BLACK_COLOR = "black";
