--- src/main/java/Pawn.java
@@ -1,12 +1,0 @@
-
-public class Pawn {
-    final private String color;
-
-    public Pawn(String color) {
-        this.color = color;
-    }
-
-    public String getColor() {
-        return color;
-    }
-}
