+++ src/test/java/PawnTest.java
@@ -1,20 +1,26 @@
 
 import org.junit.jupiter.api.*;
+import pieces.Pawn;
+
 import static org.assertj.core.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class PawnTest {
     @Test
     @DisplayName("폰이 생성되어야 한다")
     public void create() {
-        final String WHITE_COLOR = "white";
-        final String BLACK_COLOR = "black";
-
-        verifyPawn(WHITE_COLOR);
-        verifyPawn(BLACK_COLOR);
+        verifyPawn(Pawn.WHITE_COLOR);
+        verifyPawn(Pawn.BLACK_COLOR);
     }
 
     private void verifyPawn(final String color) {
         Pawn pawn = new Pawn(color);
         assertThat(pawn.getColor()).isEqualTo(color);
+    }
+
+    @Test
+    public void create_기본생성자() throws Exception {
+        Pawn pawn = new Pawn();
+        assertEquals("white", pawn.getColor());
     }
 }
